backend/requirements.txt
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
eventlet==0.33.3
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6
gevent==23.9.1
greenlet==3.0.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
PyJWT==2.8.0
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
urllib3==2.0.7
Werkzeug==3.0.0
wsproto==1.2.0
zope.event==5.0
zope.interface==6.1
